fix: run benchmark providers directly with 90s timeout budget
